+++ examples/src/index.ts
@@ -2,9 +2,13 @@
  * Main export of the @carnesen/cli-examples package
  */
 
-export { echoCommand } from './echo-command';
-export { multiplyCommand } from './multiply-command';
-export { throwErrorCommand } from './throw-error-command';
 export { advancedBranch } from './advanced-branch';
+export { demoEscapedArgumentsCommand } from './demo-escaped-arguments-command';
+export { echoCommand } from './echo-command';
+export { echoHiddenCommand } from './echo-hidden-command';
+export { echoPizzaCommand } from './echo-pizza-command';
+export { hidingBranch } from './hiding-branch';
+export { multiplyCommand } from './multiply-command';
 export { parseJsonCommand } from './parse-json-command';
+export { throwErrorCommand } from './throw-error-command';
 export { rootBranch } from './root-branch';